unsigned long is 8 bytes on 64-bit architectures
 ==13219== Conditional jump or move depends on uninitialised value(s)
 ==13219==    at 0x8058E50: LoadImage (image.c:453)
 ==13219==    by 0x80577FF: CreateIconFromFile.part.3 (icon.c:451)
 ==13219==    by 0x805B5CA: InitializeMenu (menu.c:78)
 ==13219==    by 0x805B6F2: InitializeMenu (menu.c:140)
 ==13219==    by 0x8063084: StartupRootMenu (root.c:61)
 ==13219==    by 0x804BAFE: main (main.c:534)
 ==13219==  Uninitialised value was created by a stack allocation
 ==13219==    at 0x4D07474: ??? (in /usr/lib/libpixman-1.so.0.32.4)

+++ src/image.c
@@ -444,11 +444,11 @@ ImageNode *LoadSVGImage(const char *fileName)
    g_object_unref(rh);
 
    for(i = 0; i < 4 * dim.width * dim.height; i += 4) {
-      const unsigned long temp = *(unsigned long*)&result->data[i];
-      const unsigned long alpha  = (temp >> 24) & 0xFF;
-      const unsigned long red    = (temp >> 16) & 0xFF;
-      const unsigned long green  = (temp >>  8) & 0xFF;
-      const unsigned long blue   = (temp >>  0) & 0xFF;
+      const unsigned int temp = *(unsigned int*)&result->data[i];
+      const unsigned int alpha  = (temp >> 24) & 0xFF;
+      const unsigned int red    = (temp >> 16) & 0xFF;
+      const unsigned int green  = (temp >>  8) & 0xFF;
+      const unsigned int blue   = (temp >>  0) & 0xFF;
       result->data[i + 0] = alpha;
       if(alpha > 0) {
          result->data[i + 1] = (red * 255) / alpha;
